simplify the surrogate a lot, Cliffords not quite working
