+++ src/Gates/noisechannels.jl
@@ -27,7 +27,7 @@ Apply a depolarizing noise channel to an integer Pauli string `pstr` with noise 
 Physically `p` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function apply(gate::DepolarizingNoise, pstr::PauliStringType, p, coefficient=1.0)
+function apply(gate::DepolarizingNoise, pstr::PauliStringType, p, coefficient=1.0; kwargs...)
 
     if getpauli(pstr, gate.qind) != 0   # non-identity operator
         coefficient *= (1 - p)
@@ -53,7 +53,7 @@ Apply a dephasing noise channel to an integer Pauli string `pstr` with noise str
 Physically `p` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function apply(gate::DephasingNoise, pstr::PauliStringType, p, coefficient=1.0)
+function apply(gate::DephasingNoise, pstr::PauliStringType, p, coefficient=1.0; kwargs...)
 
     pauli = getpauli(pstr, gate.qind)
     if pauli == 1 || pauli == 2   # X or Y operator
@@ -81,7 +81,7 @@ Apply a Pauli-X noise channel to an integer Pauli string `pstr` with noise stren
 Physically `p` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function apply(gate::PauliXNoise, pstr::PauliStringType, p, coefficient=1.0)
+function apply(gate::PauliXNoise, pstr::PauliStringType, p, coefficient=1.0; kwargs...)
 
     if getpauli(pstr, gate.qind) == 1   # X operator
         coefficient *= (1 - p)
@@ -107,7 +107,7 @@ Apply a Pauli-Y noise channel to an integer Pauli string `pstr` with noise stren
 Physically `p` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function apply(gate::PauliYNoise, pstr::PauliStringType, p, coefficient=1.0)
+function apply(gate::PauliYNoise, pstr::PauliStringType, p, coefficient=1.0; kwargs...)
 
     if getpauli(pstr, gate.qind) == 2   # Y operator
         coefficient *= (1 - p)
@@ -133,7 +133,7 @@ Apply a Pauli-Z noise channel to an integer Pauli string `pstr` with noise stren
 Physically `p` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function apply(gate::PauliZNoise, pstr::PauliStringType, p, coefficient=1.0)
+function apply(gate::PauliZNoise, pstr::PauliStringType, p, coefficient=1.0; kwargs...)
 
     if getpauli(pstr, gate.qind) == 3   # Z operator
         coefficient *= (1 - p)
@@ -160,7 +160,7 @@ Returns a tuple of either a single pair of Pauli string and coefficient or two p
 Physically `gamma` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function apply(gate::AmplitudeDampingNoise, pstr::PauliStringType, gamma, coefficient=1.0)
+function apply(gate::AmplitudeDampingNoise, pstr::PauliStringType, gamma, coefficient=1.0; kwargs...)
 
     if actsdiagonally(gate, pstr)  # test for Z operator
         return diagonalapply(gate, pstr, gamma, coefficient)
@@ -175,7 +175,7 @@ end
 Check if the amplitude damping noise channel acts diagonally on the Pauli string `pstr`.
 This implies no splitting (acting diagonally) which happens when acting on I, X, and Y.
 """
-function actsdiagonally(gate::AmplitudeDampingNoise, pstr::PauliStringType)
+function actsdiagonally(gate::AmplitudeDampingNoise, pstr::PauliStringType; kwargs...)
     return getpauli(pstr, gate.qind) != 3
 end
 
@@ -188,7 +188,7 @@ Returns a tuple of Pauli string and coefficient.
 Physically `gamma` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function diagonalapply(gate::AmplitudeDampingNoise, pstr::PauliStringType, gamma, coefficient=1.0)
+function diagonalapply(gate::AmplitudeDampingNoise, pstr::PauliStringType, gamma, coefficient=1.0; kwargs...)
 
     local_pauli = getpauli(pstr, gate.qind)
 
@@ -208,7 +208,7 @@ Returns a tuple of two pairs of Pauli strings and coefficients.
 Physically `gamma` is restricted to the range `[0, 1]`.
 A coefficient of the Pauli string can optionally be passed as `coefficient`.
 """
-function splitapply(gate::AmplitudeDampingNoise, pstr::PauliStringType, gamma, coefficient=1.0)
+function splitapply(gate::AmplitudeDampingNoise, pstr::PauliStringType, gamma, coefficient=1.0; kwargs...)
     new_pstr = setpauli(pstr, 0, gate.qind)
     return pstr, (1 - gamma) * coefficient, new_pstr, gamma * coefficient
 end
